+++ packages/core/tests/test_plugins.py
@@ -38,6 +38,7 @@ from prismriver_lyrics.plugins.seekalyric import SeekALyricPlugin
 from prismriver_lyrics.plugins.showmelyrics import ShowMeLyricsPlugin
 from prismriver_lyrics.plugins.snakeroot import SnakerootPlugin
 from prismriver_lyrics.plugins.song_guru import SongGuruPlugin
+from prismriver_lyrics.plugins.songlyrics import SongLyricsPlugin
 from prismriver_lyrics.plugins.songmeanings import SongMeaningsPlugin
 from prismriver_lyrics.plugins.synclrc import SyncLrcPlugin
 from prismriver_lyrics.plugins.utaten import UtaTenPlugin
@@ -477,6 +478,14 @@ class TestPlugins(PluginTestCase):
             "Сектор Газа",
             "Колхозный панк",
             "ea86197ca454df706fa1221f2125cc0c",
+        )
+
+    def test_songlyrics_01(self):
+        self.check_plugin(
+            SongLyricsPlugin(),
+            "Metallica",
+            "Sad But True",
+            "325faee229d954ac1f27c038192035d1",
         )
 
     def test_songmeanings_01(self):